requirements.txt
matplotlib~=3.10.1
pydantic~=2.11.3
crewai==0.114.0
grpcio==1.71.0
langchain-community~=0.3.22
google-generativeai>=0.3.0